+++ src/main/java/com/liuzhihang/doc/view/ui/window/MethodNode.java
@@ -15,7 +15,10 @@ import icons.DocViewIcons;
 
 import javax.swing.*;
 import java.awt.event.InputEvent;
-import java.util.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
  * 目录树上的一个节点
@@ -73,11 +76,8 @@ public class MethodNode extends DocViewNode {
     @Override
     public List<DocView> docViewList() {
         DocViewService service = DocViewService.getInstance(psiClass.getProject(), psiClass);
-        if (service != null) {
-            return Collections.singletonList(service.buildClassMethodDoc(psiClass, psiMethod));
-        }
+        return Collections.singletonList(service.buildClassMethodDoc(psiClass, psiMethod));
 
-        return new LinkedList<>();
     }
 
     @Override