+++ CHANGELOG.md
@@ -4,14 +4,19 @@
 
 ### Added
 
-### Changed
+- 增加识别 `@RequestParam("file") MultipartFile file`；fix:#52
+- 增加字段名称设置，支持 `com.fasterxml.jackson.annotation.JsonProperty` 注解自定义名称，且支持设置；fix:#72
 
-- 移除编辑文档功能，编辑文档，可以直接修改注释，或者修改生成后的文档
+### Changed
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- 重载方法异常
+- fix:#81
+- IDEA-2023.3 读取注视 ReadWriteLock 异常
 
 ### Security